requirements.txt
flatten-dict
ruamel.yaml
requests